+++ Pod/Classes/NSURLSession/MockRegister.swift
@@ -8,16 +8,28 @@
 
 import Foundation
 
-class MockRegister {
+class MockRegister<T: SessionMock> {
     
-    private var mocks: [SessionMock] = []
+    private var permanentMocks: [T] = []
+    private var ephemeralMocks: [T] = []
     
     func removeAllMocks() {
-        self.mocks.removeAll()
+        self.permanentMocks.removeAll()
+        self.ephemeralMocks.removeAll()
     }
     
-    func addMock(mock: SessionMock) {
-        self.mocks.append(mock)
+    /**
+     Adds a mock to the resgister that does not get removed after being returned
+    */
+    func addPermanentMock(mock: T) {
+        self.permanentMocks.append(mock)
+    }
+    
+    /**
+     Adds a mock to to the register that will be removed after being returned once
+    */
+    func addEphemeralMock(mock: T) {
+        self.ephemeralMocks.append(mock)
     }
     
     /**
@@ -25,13 +37,30 @@ class MockRegister {
      be mocked
      */
     func removeAllMocks(of request: NSURLRequest) {
-        self.mocks = self.mocks.filter { item in
-            return !item.matchesRequest(request)
+        self.permanentMocks = self.permanentMocks.filter {
+            return !$0.matchesRequest(request)
+        }
+        self.ephemeralMocks = self.ephemeralMocks.filter {
+            return !$0.matchesRequest(request)
         }
     }
     
+    /*
+    Returns the next mock for the given request. If the next mock is ephemeral, 
+    it also removes it from the pool of ephemeral mocks.
+    */
     func nextSessionMockForRequest(request: NSURLRequest) -> SessionMock? {
-        for mock in mocks {
+        let mocksCopy = self.ephemeralMocks
+        
+        //Ephemeral mocks have precedence over permanent mocks
+        for (index, mock) in mocksCopy.enumerate() {
+            if mock.matchesRequest(request) {
+                self.ephemeralMocks.removeAtIndex(index)
+                return mock
+            }
+        }
+        
+        for mock in self.permanentMocks {
             if mock.matchesRequest(request) {
                 return mock
             }
@@ -39,4 +68,15 @@ class MockRegister {
         
         return nil
     }
+}
+
+extension MockRegister where T: Equatable {
+
+    /**
+     Returns true if this register contains `mock` as an ephemeral mock
+     */
+    func containsEphemeralMock(mock: T) -> Bool {
+        return self.ephemeralMocks.contains(mock)
+    }
+
 }