+++ Pod/Classes/NSURLSession/NSURLSession+Mock.swift
@@ -26,7 +26,7 @@ public enum RequestDebugLevel: Int {
  An opaque reference to a request for mocking - this can be passed back in to determine the state of a mock
 */
 public struct Handle {
-    private let sessionMock: SessionMock
+    private let sessionMock: SuccessSessionMock
 }
 
 /**
@@ -37,7 +37,7 @@ public typealias BodyFunction = [String] -> NSData
 
 extension NSURLSession {
     
-    internal static let register = MockRegister()
+    internal static let register = MockRegister<SuccessSessionMock>()
     
     /**
      The next call exactly matching `request` will successfully return `body`
@@ -48,9 +48,9 @@ extension NSURLSession {
      - parameter statusCode: The status code (default=200) returned by the session data task
      - parameter delay: A artificial delay before the session data task starts to return response and data
      */
-    public class func mockSingle(request: NSURLRequest, body: NSData?, headers: [String: String] = [:], statusCode: Int = 200, delay: Double = DefaultDelay) -> Handle {
+    public class func mockNext(request: NSURLRequest, body: NSData?, headers: [String: String] = [:], statusCode: Int = 200, delay: Double = DefaultDelay) -> Handle {
         let matcher = SimpleRequestMatcher(url: request.URL!, method: request.HTTPMethod!)
-        return self.mockSingle(matcher, response: { _ in MockResponse(body: body, statusCode: statusCode, headers: headers) }, delay: delay)
+        return self.mockNext(matcher, response: { _ in MockResponse(body: body, statusCode: statusCode, headers: headers) }, delay: delay)
     }
     
     /**
@@ -77,9 +77,9 @@ extension NSURLSession {
      - parameter statusCode: The status code (default=200) returned by the session data task
      - parameter delay: A artificial delay before the session data task starts to return response and data
      */
-    public class func mockSingle(expression: String, HTTPMethod: String = "GET", body: NSData?, headers: [String: String] = [:], statusCode: Int = 200, delay: Double = DefaultDelay) throws -> Handle {
+    public class func mockNext(expression: String, HTTPMethod: String = "GET", body: NSData?, headers: [String: String] = [:], statusCode: Int = 200, delay: Double = DefaultDelay) throws -> Handle {
         let matcher = try SimpleRequestMatcher(expression: expression, method: HTTPMethod)
-        return self.mockSingle(matcher, response: { _ in return MockResponse(body: body, statusCode: statusCode, headers: headers) }, delay: delay)
+        return self.mockNext(matcher, response: { _ in return MockResponse(body: body, statusCode: statusCode, headers: headers) }, delay: delay)
     }
 
     /**
@@ -107,9 +107,9 @@ extension NSURLSession {
      - parameter delay: A artificial delay before the session data task starts to return response and data
      - parameter body: Returns data the data to be  returned by the session data task. If this returns `nil` then the didRecieveData callback won't be called.
      */
-    public class func mockSingle(expression: String, HTTPMethod: String = "GET", headers: [String: String] = [:], statusCode: Int = 200, delay: Double = DefaultDelay, body: BodyFunction) throws -> Handle {
+    public class func mockNext(expression: String, HTTPMethod: String = "GET", headers: [String: String] = [:], statusCode: Int = 200, delay: Double = DefaultDelay, body: BodyFunction) throws -> Handle {
         let matcher = try SimpleRequestMatcher(expression: expression, method: HTTPMethod)
-        return self.mockSingle(matcher, response: { (url: NSURL, extractions: [String]) in return MockResponse(body: body(extractions), statusCode: statusCode, headers: headers) }, delay: delay)
+        return self.mockNext(matcher, response: { (url: NSURL, extractions: [String]) in return MockResponse(body: body(extractions), statusCode: statusCode, headers: headers) }, delay: delay)
     }
 
     /**
@@ -148,23 +148,23 @@ extension NSURLSession {
      - paramater handle: The returned handle from a call to mock a request i.e. mockSingle(...)
      */
     public class func hasMockConsumed(handle: Handle) -> Bool {
-        return !handle.sessionMock.canMatchRequests
+        return !register.containsEphemeralMock(handle.sessionMock)
     }
 
     //MARK: Private methods
     
     // Add a request matcher to the list of mocks
-    private class func mockSingle(matcher: RequestMatcher, response: MockResponseHandler, delay: Double) -> Handle {
-        let handle = Handle(sessionMock: SingleSuccessSessionMock(matching: matcher, response: response, delay: delay))
-        self.register.addMock(handle.sessionMock)
+    private class func mockNext(matcher: RequestMatcher, response: MockResponseHandler, delay: Double) -> Handle {
+        let handle = Handle(sessionMock: SuccessSessionMock(matching: matcher, response: response, delay: delay))
+        self.register.addEphemeralMock(handle.sessionMock)
         swizzleIfNeeded()
         return handle
     }
     
     // Add a request matcher to the list of mocks
     private class func mockEvery(matcher: RequestMatcher, response: MockResponseHandler, delay: Double) {
-        let sessionMock = SuccessSessionMock(matching: matcher, response: response, delay: delay)
-        self.register.addMock(sessionMock)
+        let mock = SuccessSessionMock(matching: matcher, response: response, delay: delay)
+        self.register.addPermanentMock(mock)
         swizzleIfNeeded()
     }